Fixes data corruption when calling sync_queue::consume()
The data was returned by reference as T&, thus temporarily
stored in a local variable. When destroying the local variable
the Garbage Collector reduced the reference count of the object
by one, effectively causing data loss and, sometimes, core
corruption.

The data is now returned by copy, so the compiler decides whether
to optimize the copy by passing a reference or not.