Remove toNullTermStr from clips wrapper
The function toNullTermStr copied the contents of an std::string
to a dynamically allocated buffer as a C NULL-terminated string,
that required latter resource liberation.

The function was removed and all its dependants now rely on the
clipsstr function that returns either
    c_str() if the string is not empty
    NULL    otherwise

This change was done to prevent memory leaks and to avoid unnecessary
data copying.